+++ src/components/Splash.js
@@ -2,7 +2,6 @@ import { Button } from '@material-ui/core';
 import React from 'react'
 import LogoSvg from '../assets/Splash/logo.svg'
 import TreeSvg from '../assets/Splash/tree.svg'
-import {titleTextstyle, contentTextStyle} from '../styles/common.js'
 
 function Splash() {
     return (        
@@ -14,12 +13,12 @@ function Splash() {
             }}
             > 
             <div className='navbar-container' style={{marginLeft:'5%', marginRight:'5%', display:'flex', flexDirection:'column'}}>
+                <a href='/' rel='noopener'>
                     <Button style={{ width:"100%",justifyContent:"flex-start", borderRadius:"16px",  maxWidth:"376px",marginTop:'5%', maxHeight:"83.22px",}}>
-                        <a href='/' rel='noopener noreferrer' >
-                            <img src={LogoSvg} style={{minHeight:"83.22px",}} alt=''/>
-                            
-                        </a>
+                        <img src={LogoSvg} style={{minHeight:"83.22px",}} alt=''/>
+                        
                     </Button>
+                </a>
             </div>
             <div className='hero-body' style={{display:'flex', marginLeft:'5%', marginRight:'5%', marginTop:'120px', flexDirection:window.innerWidth < 768 ?'column':'row'}}>
                 <div className='text-container' style={{display: 'flex', flexDirection:'column', justifyContent:'flex-start', maxWidth:window.innerWidth < 768 ?'100%':'50%'}}>
@@ -38,3 +37,16 @@ function Splash() {
     )
 }
 export default Splash;
+let titleTextstyle={
+    fontFamily:"Bai Jamjuree, sans serif",
+    fontStyle: "normal",
+    fontWeight: "bold", 
+    fontSize:"72px",
+    color:"#91D873"
+}
+let contentTextStyle={
+    ...titleTextstyle,
+    fontWeight:"500",
+    color:"white",
+    fontSize:"24px"
+}
