--- src/styles/common.js
@@ -1,19 +1,0 @@
-let titleTextstyle={
-    fontFamily:"Bai Jamjuree, sans serif",
-    fontStyle: "normal",
-    fontWeight: "bold", 
-    fontSize:"72px",
-    color:"#91D873"
-}
-let contentTextStyle={
-    ...titleTextstyle,
-    fontWeight:"500",
-    color:"white",
-    fontSize:"24px"
-}
-let verticalDividedrStyle={
-    borderLeft:"1px solid #D2D7D4", 
-    marginRight:"26px", 
-    opacity:0.4
-}
-export {titleTextstyle, contentTextStyle, verticalDividedrStyle};
